Add unpkg to example
