+++ examples/embed-map/index.html
@@ -1,9 +1,6 @@
 <html>
   <style>
-     .map {
-      height: 400px;
-      width: 100%;
-     }
+     .map { height: 400px; }
     </style>
   <body>
 
@@ -11,8 +8,8 @@
     <div class="map" data-view-embed-map='{"cityCoordinates": {"lat": -35.279115, "lng": 149.126466},"coordinates": {"lat": -35.280937,"lng": 149.130009}}'></div>
     <h1>Icelab in Melbourne</h1>
     <div class="map" data-view-embed-map='{"cityCoordinates": {"lat": -37.8136, "lng": 144.9631},"coordinates": {"lat": -37.7971762, "lng": 144.9855213}}'></div>
-
     <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDObqWvYRenuA4KR-Byx-sdXiNWJ8pQHPA"></script>
-    <script src="./bundle.js"></script>
+    <script src="https://unpkg.com/viewloader@1.1.5/dist/viewloader.js"></script>
+    <script src="index.js"></script>
   </body>
 </html>
