+++ examples/embed-map/index.js
@@ -1,9 +1,21 @@
-var viewLoader = require('../../index.js');
-var embedMap = require('./embed-map');
+function initMap(el, options){
+  const map = new google.maps.Map(el, {
+    center: options.cityCoordinates,
+    zoom: 12
+  });
 
-var views = {}
-views.embedMap = function(el, props) {
-  embedMap(el, props);
+  new google.maps.Marker({
+    position: options.coordinates,
+    map: map
+  });
+}
+
+const views = {
+  embedMap: function embedMap(el, props) {
+    google.maps.event.addDomListener(window, 'load', function onLoad () {
+      initMap(el, props);
+    });
+  }
 };
 
-viewLoader.execute(views);
+viewloader.execute(views);
